requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==3.2.10
django-ckeditor==6.2.0
django-decouple==2.1
django-dotenv==1.4.2
django-js-asset==1.2.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
webencodings==0.5.1
